Crate gluon_vm

Crate which contain the virtual machine which executes gluon programs

Modules

api
The marshalling api

channel

compiler

debug

gc

internal
Internal types and functions exposed to the main gluon crate

macros
Module providing the building blocks to create macros and expand them.

primitives
Module containing functions for interacting with gluon's primitive types.

stack

thread
The thread/vm type

types

Macros

field_decl

hlist

on_array

primitive

record
Macro that creates a record that can be passed to gluon

record_no_decl

types

vm_function

Structs

Variants

Enums

Error
Representation of all possible errors that can occur when interacting with the vm crate

Type Definitions

Result
Type returned from vm functions which may fail