Module gluon_vm::api

The marshalling api

Reexports

pub use self::record::Record;

Modules

generic
Module containing types which represent generic variables in gluon's type system

record

Structs

Array
Type which represents an array

CPrimitive

Data

Function
Type which represents an function in gluon

Generic

Hole
Marker type representing a hole

OpaqueValue
Type which represents an array in gluon Type implementing both Pushable and Getable of values of V. The actual value, V is not accessible directly but is only intended to be transferred between two different threads.

Primitive

RefPrimitive

WithVM
Wrapper type which passes acts as the type T but also passes the VM to the called function

Enums

IO
Type representing gluon's IO type#[derive(Debug)]

RuntimeResult

ValueRef

Traits

FunctionType
Trait which represents a function

Getable
Trait which allows rust values to be retrieved from the virtual machine

Pushable
Trait which allows a rust value to be pushed to the virtual machine

Userdata

VmFunction
Trait which abstracts over types which can be called by being pulling the arguments it needs from the virtual machine's stack

VmType
Trait which maps a type in rust to a type in gluon

Functions

primitive

primitive_f

Type Definitions

FunctionRef
Type which represents a function reference in gluon

GluonFunction