Struct gluon_vm::api::Function

pub struct Function<T, F> where T: Deref<Target=Thread> { /* fields omitted */ }

Type which represents an function in gluon

Methods

impl<T, F> Function<T, F> where T: Deref<Target=Thread>

fn value(&self) -> Value

impl<'vm, T, R> Function<T, fn() -> R> where T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self) -> Result<R>

impl<'vm, T, A, R> Function<T, fn(A) -> R> where A: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A) -> Result<R>

impl<'vm, T, A, B, R> Function<T, fn(A, B) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B) -> Result<R>

impl<'vm, T, A, B, C, R> Function<T, fn(A, B, C) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, C: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B, C: C) -> Result<R>

impl<'vm, T, A, B, C, D, R> Function<T, fn(A, B, C, D) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, C: Pushable<'vm>, D: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B, C: C, D: D) -> Result<R>

impl<'vm, T, A, B, C, D, E, R> Function<T, fn(A, B, C, D, E) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, C: Pushable<'vm>, D: Pushable<'vm>, E: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B, C: C, D: D, E: E) -> Result<R>

impl<'vm, T, A, B, C, D, E, F, R> Function<T, fn(A, B, C, D, E, F) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, C: Pushable<'vm>, D: Pushable<'vm>, E: Pushable<'vm>, F: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B, C: C, D: D, E: E, F: F) -> Result<R>

impl<'vm, T, A, B, C, D, E, F, G, R> Function<T, fn(A, B, C, D, E, F, G) -> R> where A: Pushable<'vm>, B: Pushable<'vm>, C: Pushable<'vm>, D: Pushable<'vm>, E: Pushable<'vm>, F: Pushable<'vm>, G: Pushable<'vm>, T: Deref<Target=Thread>, R: VmType + Getable<'vm>

fn call(&'vm mut self, A: A, B: B, C: C, D: D, E: E, F: F, G: G) -> Result<R>