Trait gluon_vm::api::Getable

pub trait Getable<'vm>: Sized {
    fn from_value(vm: &'vm Thread, value: Variants) -> Option<Self>;

    unsafe fn from_value_unsafe(vm: &'vm Thread, value: Variants) -> Option<Self> { ... }
}

Trait which allows rust values to be retrieved from the virtual machine

Required Methods

fn from_value(vm: &'vm Thread, value: Variants) -> Option<Self>

Provided Methods

unsafe fn from_value_unsafe(vm: &'vm Thread, value: Variants) -> Option<Self>

unsafe version of from_value which allows references to the internal of GcPtr's to be extracted if value is rooted