Trait gluon_vm::api::Pushable

pub trait Pushable<'vm> {
    fn push(self, vm: &'vm Thread, context: &mut Context) -> Result<()>;

    fn status_push(self, vm: &'vm Thread, context: &mut Context) -> Status where Self: Sized { ... }
}

Trait which allows a rust value to be pushed to the virtual machine

Required Methods

fn push(self, vm: &'vm Thread, context: &mut Context) -> Result<()>

Pushes self to stack. If the call is successful a single element should have been added to the stack and Ok(()) should be returned. If the call is unsuccessful Status:Error should be returned and the stack should be left intact

Provided Methods

fn status_push(self, vm: &'vm Thread, context: &mut Context) -> Status where Self: Sized