Trait gluon_vm::api::VmType

pub trait VmType {
    type Type: ?Sized + Any;
    fn make_type(vm: &Thread) -> ArcType { ... }
    fn extra_args() -> VmIndex { ... }
}

Trait which maps a type in rust to a type in gluon

Associated Types

type Type: ?Sized + Any

A version of Self which implements Any allowing a TypeId to be retrieved

Provided Methods

fn make_type(vm: &Thread) -> ArcType

Creates an gluon type which maps to Self in rust

fn extra_args() -> VmIndex

How many extra arguments a function returning this type requires. Used for abstract types which when used in return position should act like they still need more arguments before they are called