Module gluon_vm::gc

Structs

Gc
A mark and sweep garbage collector.

GcPtr
A pointer to a garbage collected value.

Move
DataDef that moves its value directly into the pointer useful for sized types

WriteOnly
Pointer type which can only be written to.

Traits

DataDef
A definition of some data which may be allocated by the garbage collector.

FromPtr
Trait which creates a typed pointer from a *mut () pointer. For Sized types this is just a cast but for unsized types some more metadata must be taken from the provided D value to make it initialize correctly.

Traverseable
Trait which must be implemented on all root types which contain GcPtr A type implementing Traverseable must call traverse on each of its fields which in turn contains GcPtr