Struct gluon_vm::gc::GcPtr

pub struct GcPtr<T: ?Sized> { /* fields omitted */ }

A pointer to a garbage collected value.

It is only safe to access data through a GcPtr if the value is rooted (stored in a place where the garbage collector will find it during the mark phase).

Methods

impl<T: ?Sized> GcPtr<T>

unsafe fn as_mut(&mut self) -> &mut T

Unsafe as it is up to the caller to ensure that this pointer is not referenced somewhere else

unsafe fn from_raw(ptr: *const T) -> GcPtr<T>

Unsafe as ptr must have been allocted by this garbage collector

fn generation(&self) -> usize

impl<'a, T: Traverseable + Send + Sync + 'a> GcPtr<T>

fn as_traverseable(self) -> GcPtr<Traverseable + Send + Sync + 'a>

Coerces self to a Traverseable trait object

impl GcPtr<str>

fn as_traverseable_string(self) -> GcPtr<Traverseable + Send + Sync>

Coerces self to a Traverseable trait object