Struct gluon_vm::gc::WriteOnly

pub struct WriteOnly<'s, T: ?Sized + 's>(_, _);

Pointer type which can only be written to.

Methods

impl<'s, T: ?Sized> WriteOnly<'s, T>

fn as_mut_ptr(&mut self) -> *mut T

Retrieves the pointer allowing it to be manipulated freely. As it points to uninitialized data care must be taken so to not read it before it has been initialized

impl<'s, T> WriteOnly<'s, T>

fn write(self, t: T) -> &'s mut T

Writes a value to the pointer and returns a pointer to the now initialized value.

impl<'s, T: Copy> WriteOnly<'s, [T]>

fn write_slice(self, s: &[T]) -> &'s mut [T]

impl<'s> WriteOnly<'s, str>

fn write_str(self, s: &str) -> &'s mut str