Trait gluon_vm::gc::Traverseable

pub trait Traverseable {
    fn traverse(&self, gc: &mut Gc) { ... }
}

Trait which must be implemented on all root types which contain GcPtr A type implementing Traverseable must call traverse on each of its fields which in turn contains GcPtr

Provided Methods

fn traverse(&self, gc: &mut Gc)