Struct gluon_vm::Variants

pub struct Variants<'a>(_);

Methods

impl<'a> Variants<'a>

unsafe fn new(value: &Value) -> Variants

Creates a new Variants value which assumes that value is alive for the lifetime of the value

fn as_ref(&self) -> ValueRef

Returns an instance of ValueRef which allows users to safely retrieve the interals of a value