Module gluon_vm::thread

The thread/vm type

Reexports

pub use gc::Traverseable;

Structs

Context

OwnedContext

Root
A rooted userdata value

RootStr
A rooted string

RootedThread
An instance of Thread which is rooted. See the Thread type for documentation on interacting with the type.

RootedValue
A rooted value

Thread
Representation of the virtual machine

Enums

Status
Enum signaling a successful or unsuccess ful call to an extern function. If an error occured the error message is expected to be on the top of the stack.

Traits

ThreadInternal
Internal functions for interacting with threads. These functions should be considered both unsafe and unstable.

Functions

alloc

Type Definitions

HookFn