Trait gluon_vm::thread::ThreadInternal

pub trait ThreadInternal {
    fn context(&self) -> OwnedContext;
    fn root<'vm, T: Userdata>(&'vm self, v: GcPtr<Box<Userdata>>) -> Option<Root<'vm, T>>;
    fn root_string<'vm>(&'vm self, ptr: GcPtr<Str>) -> RootStr<'vm>;
    fn root_value(&self, value: Value) -> RootedValue<RootedThread>;
    fn root_value_ref(&self, value: Value) -> RootedValue<&Thread>;
    fn add_bytecode(&self, name: &str, typ: ArcType, args: VmIndex, instructions: Vec<Instruction>) -> Result<()>;
    fn call_thunk(&self, closure: GcPtr<ClosureData>) -> Result<Value>;
    fn execute_io(&self, value: Value) -> Result<Value>;
    fn call_function<'b>(&'b self, stack: OwnedContext<'b>, args: VmIndex) -> Result<Option<OwnedContext<'b>>>;
    fn resume(&self) -> Result<()>;
    fn global_env(&self) -> &Arc<GlobalVmState>;
    fn set_global(&self, name: Symbol, typ: ArcType, metadata: Metadata, value: Value) -> Result<()>;
    fn deep_clone_value(&self, value: Value) -> Result<Value>;
}

Internal functions for interacting with threads. These functions should be considered both unsafe and unstable.

Required Methods

fn context(&self) -> OwnedContext

Locks and retrives this threads stack

fn root<'vm, T: Userdata>(&'vm self, v: GcPtr<Box<Userdata>>) -> Option<Root<'vm, T>>

Roots a userdata

fn root_string<'vm>(&'vm self, ptr: GcPtr<Str>) -> RootStr<'vm>

Roots a string

fn root_value(&self, value: Value) -> RootedValue<RootedThread>

Roots a value

fn root_value_ref(&self, value: Value) -> RootedValue<&Thread>

Roots a value

fn add_bytecode(&self, name: &str, typ: ArcType, args: VmIndex, instructions: Vec<Instruction>) -> Result<()>

fn call_thunk(&self, closure: GcPtr<ClosureData>) -> Result<Value>

Evaluates a zero argument function (a thunk)

fn execute_io(&self, value: Value) -> Result<Value>

Executes an IO action

fn call_function<'b>(&'b self, stack: OwnedContext<'b>, args: VmIndex) -> Result<Option<OwnedContext<'b>>>

Calls a function on the stack. When this function is called it is expected that the function exists at stack.len() - args - 1 and that the arguments are of the correct type

fn resume(&self) -> Result<()>

fn global_env(&self) -> &Arc<GlobalVmState>

fn set_global(&self, name: Symbol, typ: ArcType, metadata: Metadata, value: Value) -> Result<()>

fn deep_clone_value(&self, value: Value) -> Result<Value>