Trait gluon_vm::api::Pushable

pub trait Pushable<'vm> {
    fn push(self, vm: &'vm Thread, stack: &mut Stack) -> Status;
}

Trait which allows a rust value to be pushed to the virtual machine

Required Methods

fn push(self, vm: &'vm Thread, stack: &mut Stack) -> Status

Pushes self to stack. If the call is successful a single element should have been added to the stack and Status::Ok should be returned. If the call is unsuccessful Status:Error should be returned and the stack should be left intact